requirements/base.txt
torch>=1.9.0
torchvision>=0.10.0
efficientnet-pytorch>=0.7.1
albumentations>=1.0.0
numpy>=1.19.2
pandas>=1.3.0
Pillow>=8.3.0
scikit-learn>=0.24.2
tqdm>=4.62.0
PyYAML>=5.4.1
tensorboard>=2.6.0
